LayoutTests/media/video-src-blob-using-open-panel-expected.txt

OPEN FILE PANEL
This tests the ability of the <video> element to load blob URLs. In the browser, select a video file:

EVENT(change)
EVENT(loadedmetadata)
END OF TEST
